renamed src, to etnn
