+++ etnn/legacy/simplicial_data/rips_lift.py
